requirements.txt
liboqs-python
dnspython